Issue #ISAICP-2924: Make shared directory optional.

+++ src/Phing/ImportRdfFixtures.php
@@ -87,37 +87,41 @@ class ImportRdfFixtures extends \Task {
   public function main() {
     $path_parts = explode('/', __DIR__);
     $base_dir = implode('/', array_slice($path_parts, 0, -2));
-    $fixtures_path = $base_dir . '/resources/fixtures';
-    if (substr($this->sharedDirectory, -1) !== '/') {
-      $this->sharedDirectory .= '/';
-    }
-    // When the database server and the web server are two distinct machines,
-    // we need to proxy the rdf files from one machine to the other, by using a
-    // shared mount.
-    // Clean directory from previous deploy.
-    if (is_dir($this->sharedDirectory)) {
-      array_map('unlink', glob($this->sharedDirectory . '*'));
-      rmdir($this->sharedDirectory);
-    }
+    $fixtures_path = $base_dir . '/resources/fixtures/';
 
-    $dir = opendir($fixtures_path);
-    mkdir($this->sharedDirectory, 0777, TRUE);
-    while (($file = readdir($dir)) !== FALSE) {
-      if (
-        ($file === '.') ||
-        ($file === '..') ||
-        (is_dir($fixtures_path . '/' . $file))
-      ) {
-        continue;
+    // When shared directory is in use, override fixture path.
+    if (!empty($this->sharedDirectory)) {
+      if (substr($this->sharedDirectory, -1) !== '/') {
+        $this->sharedDirectory .= '/';
       }
-      copy($fixtures_path . '/' . $file, $this->sharedDirectory . '/' . $file);
-    }
-    closedir($dir);
+      // When the database server and the web server are two distinct machines,
+      // we need to proxy the rdf files from one machine to the other, by using a
+      // shared mount.
+      // Clean directory from previous deploy.
+      if (is_dir($this->sharedDirectory)) {
+        array_map('unlink', glob($this->sharedDirectory . '*'));
+        rmdir($this->sharedDirectory);
+      }
 
+      $dir = opendir($fixtures_path);
+      mkdir($this->sharedDirectory, 0777, TRUE);
+      while (($file = readdir($dir)) !== FALSE) {
+        if (
+          ($file === '.') ||
+          ($file === '..') ||
+          (is_dir($fixtures_path . $file))
+        ) {
+          continue;
+        }
+        copy($fixtures_path . $file, $this->sharedDirectory . $file);
+      }
+      closedir($dir);
+      $fixtures_path = $this->sharedDirectory;
+    }
 
     // Reset the import table (Needed for re-import).
     $this->execute("delete from db.dba.load_list;");
-    foreach (glob($this->sharedDirectory . '*.rdf') as $rdf_file_path) {
+    foreach (glob($fixtures_path . '*.rdf') as $rdf_file_path) {
       $filename = array_pop(explode('/', $rdf_file_path));
       $file = str_replace('.rdf', '', $filename);
       $graph_name = 'http://' . strtolower($file);
